+++ src/index.ts
@@ -1,6 +1,7 @@
+export { Snowflake } from 'discord-api-types/globals';
 export * from './constants';
 export * from './http';
 export * from './gateway';
 export * from './exceptions';
 export * from './clients';
-export * from './utils';
+export * as utils from './utils';
