+++ src/structures/channel/voice/StageChannel.ts
@@ -1,4 +1,4 @@
-import { APIChannel } from 'discord-api-types';
+import type { APIChannel } from 'discord-api-types/v8';
 import { VoiceBasedChannel } from './VoiceBasedChannel';
 
 export class StageChannel extends VoiceBasedChannel {
