+++ src/structures/emoji/ReactionEmoji.ts
@@ -1,5 +1,5 @@
-import { APIPartialEmoji } from 'discord-api-types';
-import { MessageReaction } from '../message/MessageReaction';
+import type { APIPartialEmoji } from 'discord-api-types/v8';
+import type { MessageReaction } from '../message/MessageReaction';
 import { Emoji } from './Emoji';
 
 export class ReactionEmoji extends Emoji {
